Use framework version 1.0.11
