+++ src/ios/ProximityKit.framework/Headers/ProximityKit.h
@@ -1,0 +1,18 @@
+//
+//  Proximity Kit
+//
+//  Copyright (c) 2013 Radius Networks. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <ProximityKit/RPKManager.h>
+#import <ProximityKit/RPKBeacon.h>
+#import <ProximityKit/RPKBeaconRegion.h>
+#import <ProximityKit/RPKKit.h>
+#import <ProximityKit/RPKLogger.h>
+#import <ProximityKit/RPKManagerDelegate.h>
+#import <ProximityKit/RPKMap.h>
+#import <ProximityKit/RPKRegion.h>
+#import <ProximityKit/RPKManager.h>
+#import <ProximityKit/RPKCircle.h>
+#import <ProximityKit/RPKAnalyticsEvent.h>
