--- src/ios/ProximityKit.framework/Versions/A/Headers/ProximityKit.h
@@ -1,8 +1,0 @@
-//
-//  Proximity Kit
-//
-//  Copyright (c) 2013 Radius Networks. All rights reserved.
-//
-
-#import <Foundation/Foundation.h>
-#import <ProximityKit/RPKManager.h>
