--- src/ios/ProximityKit.framework/Versions/A/Headers/RPKCircle.h
@@ -1,36 +1,0 @@
-//  Copyright (c) 2013 Radius Networks. All rights reserved.
-
-#import <Foundation/Foundation.h>
-#import <CoreLocation/CoreLocation.h>
-#import <CoreLocation/CLCircularRegion.h>
-#import "RPKRegion.h"
-
-/*!
- * @interface PKCircle
- * @discussion Contains a circle region defined by GPS coordinates with a
- *             latitude and longitude.
- */
-@interface RPKCircle : RPKRegion
-
-/*!
- * @property latitude
- * @discussion The latitude value for the center GPS coordinates
- */
-@property (readonly) float latitude;
-
-/*!
- * @property longitude
- * @discussion The longitude value for the center GPS coordinates
- */
-@property (readonly) float longitude;
-
-/*!
- * @property radius
- * @discussion The length of the radius in meters
- */
-@property (readonly) float radius;
-
-- (id)initWith:(NSDictionary *)dict;
-- (CLRegion *)region;
-
-@end
