--- src/ios/ProximityKit.framework/Versions/A/Resources/Info.plist
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>CFBundleName</key>
-	<string>ProximityKit</string>
-	<key>CFBundleIdentifier</key>
-	<string>com.radiusnetworks.ProximityKit</string>
-	<key>CFBundlePackageType</key>
-	<string>FMWK</string>
-	<key>CFBundleSignature</key>
-	<string>????</string>
-	<key>CFBundleVersion</key>
-	<string>0.4.1</string>
-	<key>CFBundleShortVersionString</key>
-	<string>0.4.1</string>
-	<key>NSHumanReadableCopyright</key>
-	<string>Copyright © 2014 Radius Networks</string>
-	<key>CFBundleGetInfoString</key>
-	<string>Proximity Kit Framework</string>
-</dict>
-</plist>
