+++ src/ios/ProximityKit.framework/_CodeSignature/CodeResources
@@ -1,0 +1,207 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>files</key>
+	<dict>
+		<key>Headers/ProximityKit.h</key>
+		<data>
+		s482bu3VorzUBjjgEGB0eeyPxNU=
+		</data>
+		<key>Headers/RPKAnalyticsEvent.h</key>
+		<data>
+		i5uylpvuXDv1nNTgqzrYf7TLAVw=
+		</data>
+		<key>Headers/RPKBeacon.h</key>
+		<data>
+		/TkgH9VaQJGA3MN6JuidSAwL18A=
+		</data>
+		<key>Headers/RPKBeaconRegion.h</key>
+		<data>
+		o+TOuk7DXsEkactik/rBGe4Z/e4=
+		</data>
+		<key>Headers/RPKCircle.h</key>
+		<data>
+		Yx+CW7vR6RLlVv/Yylz/dGgB3AM=
+		</data>
+		<key>Headers/RPKKit.h</key>
+		<data>
+		Xpvn231JQpEemoZpg1273NzlJ8s=
+		</data>
+		<key>Headers/RPKLogger.h</key>
+		<data>
+		ev4wqPW0DfvmanEaMkVEru+tC7M=
+		</data>
+		<key>Headers/RPKManager.h</key>
+		<data>
+		bB419RdqzDsj1IBRjV+6ivESQbY=
+		</data>
+		<key>Headers/RPKManagerDelegate.h</key>
+		<data>
+		t4p9S4hBTW/MaHwMMSN/aCHkIeY=
+		</data>
+		<key>Headers/RPKMap.h</key>
+		<data>
+		AiQmCdDlHi3TBbwqszT5H2nR9II=
+		</data>
+		<key>Headers/RPKRegion.h</key>
+		<data>
+		Igcrv+8diJJMiTcyzKcJt/WBdHw=
+		</data>
+		<key>Info.plist</key>
+		<data>
+		28NMfwj8qNExdpNYiQpEW80vNPM=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		QEXKhVpsHZxA8Fb08ZmjBK3rHcc=
+		</data>
+	</dict>
+	<key>files2</key>
+	<dict>
+		<key>Headers/ProximityKit.h</key>
+		<data>
+		s482bu3VorzUBjjgEGB0eeyPxNU=
+		</data>
+		<key>Headers/RPKAnalyticsEvent.h</key>
+		<data>
+		i5uylpvuXDv1nNTgqzrYf7TLAVw=
+		</data>
+		<key>Headers/RPKBeacon.h</key>
+		<data>
+		/TkgH9VaQJGA3MN6JuidSAwL18A=
+		</data>
+		<key>Headers/RPKBeaconRegion.h</key>
+		<data>
+		o+TOuk7DXsEkactik/rBGe4Z/e4=
+		</data>
+		<key>Headers/RPKCircle.h</key>
+		<data>
+		Yx+CW7vR6RLlVv/Yylz/dGgB3AM=
+		</data>
+		<key>Headers/RPKKit.h</key>
+		<data>
+		Xpvn231JQpEemoZpg1273NzlJ8s=
+		</data>
+		<key>Headers/RPKLogger.h</key>
+		<data>
+		ev4wqPW0DfvmanEaMkVEru+tC7M=
+		</data>
+		<key>Headers/RPKManager.h</key>
+		<data>
+		bB419RdqzDsj1IBRjV+6ivESQbY=
+		</data>
+		<key>Headers/RPKManagerDelegate.h</key>
+		<data>
+		t4p9S4hBTW/MaHwMMSN/aCHkIeY=
+		</data>
+		<key>Headers/RPKMap.h</key>
+		<data>
+		AiQmCdDlHi3TBbwqszT5H2nR9II=
+		</data>
+		<key>Headers/RPKRegion.h</key>
+		<data>
+		Igcrv+8diJJMiTcyzKcJt/WBdHw=
+		</data>
+		<key>Modules/module.modulemap</key>
+		<data>
+		QEXKhVpsHZxA8Fb08ZmjBK3rHcc=
+		</data>
+	</dict>
+	<key>rules</key>
+	<dict>
+		<key>^</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^version.plist$</key>
+		<true/>
+	</dict>
+	<key>rules2</key>
+	<dict>
+		<key>.*\.dSYM($|/)</key>
+		<dict>
+			<key>weight</key>
+			<real>11</real>
+		</dict>
+		<key>^</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^(.*/)?\.DS_Store$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>2000</real>
+		</dict>
+		<key>^(Frameworks|SharedFrameworks|PlugIns|Plug-ins|XPCServices|Helpers|MacOS|Library/(Automator|Spotlight|LoginItems))/</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^.*</key>
+		<true/>
+		<key>^.*\.lproj/</key>
+		<dict>
+			<key>optional</key>
+			<true/>
+			<key>weight</key>
+			<real>1000</real>
+		</dict>
+		<key>^.*\.lproj/locversion.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>1100</real>
+		</dict>
+		<key>^Info\.plist$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^PkgInfo$</key>
+		<dict>
+			<key>omit</key>
+			<true/>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^[^/]+$</key>
+		<dict>
+			<key>nested</key>
+			<true/>
+			<key>weight</key>
+			<real>10</real>
+		</dict>
+		<key>^embedded\.provisionprofile$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+		<key>^version\.plist$</key>
+		<dict>
+			<key>weight</key>
+			<real>20</real>
+		</dict>
+	</dict>
+</dict>
+</plist>
